2.0. audio: Audio pipeline is now in 32 bit.      decklink_consumer: Added support for 32 bit audio.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.0.2@1321 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/mixer/audio/audio_mixer.h
@@ -40,7 +40,7 @@ public:
 	virtual void visit(core::write_frame& frame);
 	virtual void end();
 
-	std::vector<int16_t> mix();
+	std::vector<int32_t> mix();
 	
 	audio_mixer& operator=(audio_mixer&& other);
 private:
