+++ core/mixer/read_frame.h
@@ -37,10 +37,10 @@ class read_frame : boost::noncopyable
 {
 public:
 	read_frame();
-	read_frame(ogl_device& ogl, size_t size, safe_ptr<host_buffer>&& image_data, std::vector<int16_t>&& audio_data);
+	read_frame(ogl_device& ogl, size_t size, safe_ptr<host_buffer>&& image_data, std::vector<int32_t>&& audio_data);
 
 	virtual const boost::iterator_range<const uint8_t*> image_data();
-	virtual const boost::iterator_range<const int16_t*> audio_data();
+	virtual const boost::iterator_range<const int32_t*> audio_data();
 
 	virtual size_t image_size() const;
